+++ src/graphql/gqls/consentManager.ts
@@ -7,7 +7,6 @@ export const PURPOSES = gql`
       purposes {
         id
         name
-        trackingType
       }
     }
   }
@@ -34,11 +33,9 @@ export const EXPERIENCES = gql`
         viewState
         purposes {
           name
-          trackingType
         }
         optedOutPurposes {
           name
-          trackingType
         }
         browserLanguages
         browserTimeZones
@@ -328,10 +325,10 @@ export const UPDATE_CONSENT_MANAGER_DOMAINS = gql`
 export const UPDATE_CONSENT_MANAGER_PARTITION = gql`
   mutation TranscendCliUpdateConsentManagerPartition(
     $airgapBundleId: ID!
-    $partition: String!
+    $partitionId: ID!
   ) {
     updateConsentManagerPartition(
-      input: { id: $airgapBundleId, partition: $partition }
+      input: { id: $airgapBundleId, partitionId: $partitionId }
     ) {
       clientMutationId
     }